feat: 現在時刻表示をFormationタブから全画面共通に移動
タブレイアウトに現在時刻バッジを配置し、全タブ画面で表示されるように変更。
タブバーの高さをSafeAreaのinsetsから動的に計算して正しい位置に配置。
